Remove growth fit curve upon length data changed
If the fit growth data matches the current gatherFishAgeLengthData()
in growthFrequentistFitBoot() then we plot the fit and CI. If the
length (age) data is updated then we return a NULL object from the
reactive expression leading to geom_blanks for fit and conf interval.
Change slider curve colour to grey in growth plot to be less noticeable.