Change labels for assessment radio and action button
Text for radio button is changed to "Assessment - temporal basis" and
options are "all periods" or "annual" if yearly data is available. We
also change the text on the model fit button to match the selected
method, e.g. apply LB-SPR or apply LIME.

Code modifications
Change ui element (radioButton) name from
	visualiseLengthComposition
to
	analyseLengthComposition

Change if statements in plotResponsiveLengthComposition, plotLBAModelFit,
binLengthData so that input$analyseLengthComposition is compared to
options "all periods" or "annual".

Insert observe statement to updateActionButton label according to
input$lengthBasedAssessmentMethod.