Adjust server for single column of length data
Change catchdata_table so it is an eventReactive dependent on
input$selectCols and not input$checkboxCatchData (less flexible but
allows a single column of data.

Fix anyAgeData reactive so it can handle NULL input$checkboxCatchData.
Change exploreLengthAtAge reactive so it provides button to move to
Growth page even when there is no age data.