Invoke local LBSPRDome package for LBSPR functions
In optLBSPRDome, all elements of the variance-covariance matrix are
set to NA when there are problems with Hessian matrix (e.g. not
positive semi-definite or singular). In server code, we indicate
unreliable optimisations in plotFishingEstimateOutput reactive by
specifying pch values (1, for unreliable, 19 for reliable standard
error) in time history plot of fishing estimates.